SINGLE_DASH_IS_MULTIFLAG supports trailing param
Enhance the special case where the last of combined single-character
options is registered as a parameter option. E.g.:

    p.add_param("f");

And then parsing "-cvf" will treat the `-f` as a parameter option,
instead of as a flag.

Note: This is consistent with the existing special case of _just_ `-f`,
which was already treated as a parameter option.

Implements #28